Use the height/width of the heightmap to render it.
Previously, I was erroneously using the height/width of the window to
decide how many triangles I would use to render the heightmap. This
slowed down rendering, and introduced UB, as these values were undefined
in the VBO. Remedied this, and improved my naming convention to avoid
such a mistake in the future.